requirements.txt
asgiref==3.9.1
Django==5.2.6
django-extensions==4.1
django-silk==5.4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gprof2dot==2025.4.14
pillow==11.3.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
